[native_assets_cli] Split up syntax generator
