+++ pkgs/json_syntax_generator/lib/src/generator/helper_library.dart
@@ -41,6 +41,15 @@ extension on Map<String, Object?> {
     return list.cast();
   }
 
+  List<T>? optionalListParsed<T extends Object?>(
+    String key,
+    T Function(Object?) elementParser,
+  ) {
+    final jsonValue = optionalList(key);
+    if (jsonValue == null) return null;
+    return [for (final element in jsonValue) elementParser(element)];
+  }
+
   Map<String, T> map$<T extends Object?>(String key) =>
       _castMap<T>(get<Map<String, Object?>>(key), key);
 
@@ -91,6 +100,14 @@ extension on Map<String, Object?> {
       return Uri.directory(path);
     }
     return Uri.file(path);
+  }
+
+  void setOrRemove(String key, Object? value) {
+    if (value == null) {
+      remove(key);
+    } else {
+      this[key] = value;
+    }
   }
 }
